refactor: extract shared debug sub-components (#PAT-099)
Aligns frontend with the PAT-098 backend ExecutionErrorInfo
consolidation. The two debug panels (editor DebugPanel and
CdsDebugPanel) previously each re-implemented trace tables and the
ELM JSON viewer, and only the CDS panel had a structured error
alert — the editor had none.

Extracts 4 shared components under components/debug/:
- ExpressionTraceTable
- RetrieveTraceTable
- ElmJsonViewer
- ExecutionErrorAlert

DebugPanel shrinks from 247 → ~30 lines, delegating to the shared
tables. CdsDebugPanel keeps CDS-only sections (prefetch, FHIR
diagnostics, invocation context, dryRun) but uses the shared
ExecutionErrorAlert and trace tables rather than embedding the
editor's DebugPanel.

Frontend type CdsErrorInfo renamed to ExecutionErrorInfo to match
backend. Deprecated type alias kept for existing imports.

Shared debug i18n strings moved from editor.debug.* and cds.debug.*
to common.debug.* (both en and zh-TW), including two new phase
labels (fhir_retrieval, population_evaluation) introduced by the
backend's unified ErrorPhase enum. Phase→severity mapping expanded
to cover these plus card_generation.

New unit test: ExecutionErrorAlert.test.tsx (4 tests). tsc --noEmit
clean. Full vitest run deferred to CI — local Windows hits EMFILE
on @mui/icons-material barrel imports, unrelated to this refactor.

Refs: #325
